# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem pencatatan absensi sholat siswa",
        "title": "API Sistem Absensi Sholat",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/history/siswa": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil riwayat absensi sholat untuk siswa yang sedang login. Dapat dilihat per minggu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Ambil riwayat absensi siswa (per minggu)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Minggu ke berapa (0 = minggu ini, 1 = minggu lalu, dst). Default: 0",
                        "name": "week",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat absensi berhasil diambil",
                        "schema": {
                            "$ref": "#/definitions/handlers.HistorySiswaResponse"
                        }
                    },
                    "401": {
                        "description": "Tidak terotentikasi",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Siswa tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/history/staff": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil riwayat absensi semua siswa dengan filter. Hanya untuk admin, guru, dan wali kelas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Ambil riwayat absensi untuk staff (admin/guru/wali_kelas)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tanggal mulai (format: YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tanggal akhir (format: YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan kelas",
                        "name": "kelas",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan jurusan",
                        "name": "jurusan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan NIS siswa tertentu",
                        "name": "nis",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter berdasarkan status (hadir/izin/sakit/alpha)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nomor halaman (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah per halaman (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Riwayat absensi berhasil diambil",
                        "schema": {
                            "$ref": "#/definitions/handlers.HistoryStaffResponse"
                        }
                    },
                    "401": {
                        "description": "Tidak terotentikasi",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Akses ditolak",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Kesalahan server internal",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.AbsensiHistoryItem": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "hari": {
                    "type": "string"
                },
                "id_absen": {
                    "type": "integer"
                },
                "jenis_sholat": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                }
            }
        },
...
        "handlers.AbsensiStaffItem": {
            "type": "object",
            "properties": {
                "deskripsi": {
                    "type": "string"
                },
                "hari": {
                    "type": "string"
                },
                "id_absen": {
                    "type": "integer"
                },
                "jenis_sholat": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                },
                "nama_siswa": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.HistoryFilters": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "jurusan": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.HistorySiswaData": {
            "type": "object",
            "properties": {
                "absensi": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.AbsensiHistoryItem"
                    }
                },
                "end_date": {
                    "type": "string"
                },
                "periode": {
                    "type": "string"
                },
                "siswa": {
                    "$ref": "#/definitions/handlers.SiswaInfo"
                },
                "start_date": {
                    "type": "string"
                },
                "statistik": {
                    "$ref": "#/definitions/handlers.HistoryStatistik"
                }
            }
        },
        "handlers.HistorySiswaResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.HistorySiswaData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.HistoryStaffData": {
            "type": "object",
            "properties": {
                "absensi": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.AbsensiStaffItem"
                    }
                },
                "filters": {
                    "$ref": "#/definitions/handlers.HistoryFilters"
                },
                "pagination": {
                    "$ref": "#/definitions/handlers.PaginationInfo"
                },
                "statistik": {
                    "$ref": "#/definitions/handlers.LaporanStatistik"
                }
            }
        },
        "handlers.HistoryStaffResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handlers.HistoryStaffData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.HistoryStatistik": {
            "type": "object",
            "properties": {
                "persentase_kehadiran": {
                    "type": "number"
                },
                "total_absensi": {
                    "type": "integer"
                },
                "total_alpha": {
                    "type": "integer"
                },
                "total_hadir": {
                    "type": "integer"
                },
                "total_izin": {
                    "type": "integer"
                },
                "total_sakit": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.LaporanStatistik": {
            "type": "object",
            "properties": {
                "persentase_alpha": {
                    "type": "number"
                },
                "persentase_hadir": {
                    "type": "number"
                },
                "persentase_izin": {
                    "type": "number"
                },
                "persentase_sakit": {
                    "type": "number"
                },
                "rata_rata_kehadiran": {
                    "type": "number"
                },
                "total_absensi": {
                    "type": "integer"
                },
                "total_alpha": {
                    "type": "integer"
                },
                "total_hadir": {
                    "type": "integer"
                },
                "total_izin": {
                    "type": "integer"
                },
                "total_sakit": {
                    "type": "integer"
                },
                "total_siswa": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.PaginationInfo": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.SiswaInfo": {
            "type": "object",
            "properties": {
                "jurusan": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                },
                "nama_siswa": {
                    "type": "string"
                },
                "nis": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}